Don't rely on `flymake` internals. Fix compatibility with edge Emacs
`flymake' bundled with the edge/master branch version of Emacs no
longer provides the `flymake--overlays' function[^1], which breaks the
core functionality of this package.

Even though overlays are now available via
`flymake--really-all-overlays' (as seen in the linked commit), I
thought it was best to stick to the public API functions. I don't
really know elisp, so hopefully my assumption that functions with `--`
in their name are internal (i.e. not part of the public API) is
correct.

[^1]: https://github.com/emacs-mirror/emacs/commit/5792ea14ad69ae1ed5584dff8c9b7d5ee702aaed